コンポーネント描画，オフスクリーン描画時の頂点インデックスを、uint32からuint16に変更
デモ用の、メッシュを多用した非常に大きなデータでも、頂点数１万程度だった。
現実的には65535の範囲内で十分なハズ。

+++ Ss6PlayerExamples/Plugins/SpriteStudio6/Source/SpriteStudio6/Private/Component/SsPlayerComponent.cpp
@@ -345,7 +345,7 @@ void USsPlayerComponent::SendRenderDynamicData_Concurrent()
 		case ESsPlayerComponentRenderMode::Default:
 			{
 				TArray<FSsRenderPartsProxy::FSsPartVertex> RenderVertices;
-				TArray<uint32> RenderIndices;
+				TArray<uint16> RenderIndices;
 				TArray<FSsRenderPartsProxy::FSsPartPrimitive> RenderPrimitives;
 				{
 					const TArray<FSsRenderPart>& RenderParts = Player.GetRenderParts();
@@ -537,7 +537,7 @@ void USsPlayerComponent::SendRenderDynamicData_Concurrent()
 					FSendSsRenderData,
 					FSsRenderPartsProxy*, SsPartsProxy, (FSsRenderPartsProxy*)SceneProxy,
 					TArray<FSsRenderPartsProxy::FSsPartVertex>, InRenderVertices, RenderVertices,
-					TArray<uint32>, InRenderIndices, RenderIndices,
+					TArray<uint16>, InRenderIndices, RenderIndices,
 					TArray<FSsRenderPartsProxy::FSsPartPrimitive>, InRenderPrimitives, RenderPrimitives,
 				{
 						SsPartsProxy->SetDynamicData_RenderThread(InRenderVertices, InRenderIndices, InRenderPrimitives);
