+++ Ss6PlayerExamples/Plugins/SpriteStudio6/Source/SpriteStudio6/Public/Data/SsTypes.h
@@ -769,7 +769,7 @@ struct FSsRenderMeshVertex
 struct FSsRenderMesh
 {
 	TArray<FSsRenderMeshVertex> Vertices;
-	TArray<uint32> Indices;
+	TArray<uint16> Indices;
 	FColor Color;
 	float ColorBlendRate;
 };
